# lib/api-zod/src/generated/types/healthStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Deterministic local Technocore protocol conformance and bounded chaos lab
 * OpenAPI spec version: 0.9.1
 */

export interface HealthStatus {
  status: string;
}