# lib/api-zod/src/generated/types/runCase.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Deterministic local Technocore protocol conformance and bounded chaos lab
 * OpenAPI spec version: 0.9.1
 */
import type { ByteDiff } from './byteDiff';
import type { Evidence } from './evidence';
import type { RunCaseInput } from './runCaseInput';
import type { RunCaseSeverity } from './runCaseSeverity';
import type { RunCaseStatus } from './runCaseStatus';

export interface RunCase {
  id: string;
  title: string;
  category: string;
  severity: RunCaseSeverity;
  status: RunCaseStatus;
  implementationId?: string;
  input: RunCaseInput;
  expected: unknown;
  actual: unknown;
  /** @nullable */
  expectedCanonical: string | null;
  /** @nullable */
  actualCanonical: string | null;
  byteDiff: ByteDiff | null;
  /** @minimum 0 */
  durationMs: number;
  citation: string;
  evidence: Evidence[];
}