# lib/api-zod/src/generated/types/runCounts.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Deterministic local Technocore protocol conformance and bounded chaos lab
 * OpenAPI spec version: 0.9.1
 */

export interface RunCounts {
  total: number;
  passed: number;
  failed: number;
  unsupported?: number;
  errors: number;
}